iam/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devault-iam"
version = "0.1.0"
description = "Standalone Identity & Access service for DeVault (humans + control-plane API keys)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.14.0",
  "redis>=5.2.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.28.0",
  "PyJWT[crypto]>=2.9.0",
  "cryptography>=43.0.0",
  "argon2-cffi>=23.1.0",
  "pyotp>=2.9.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.0.0",
  "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.28.0",
  "fakeredis>=2.23.0",
]

[project.scripts]
devault-iam-serve = "devault_iam.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
